+++ tests/im/test_service.py
@@ -25,19 +25,15 @@ import aioxmpp.im.service as im_service
 
 from aioxmpp.testutils import (
     make_connected_client,
+    make_listener,
 )
 
 
 class TestConversationService(unittest.TestCase):
     def setUp(self):
-        self.listener = unittest.mock.Mock()
         self.cc = make_connected_client()
         self.s = im_service.ConversationService(self.cc)
-
-        for ev in ["on_conversation_added"]:
-            handler = getattr(self.listener, ev)
-            handler.return_value = None
-            getattr(self.s, ev).connect(handler)
+        self.listener = make_listener(self.s)
 
     def tearDown(self):
         del self.s
@@ -49,13 +45,16 @@ class TestConversationService(unittest.TestCase):
             [],
         )
 
-    def test__add_conversation(self):
+    def test__add_conversation_emits_on_conversation_added(self):
         conv = unittest.mock.Mock()
         self.s._add_conversation(conv)
-        self.listener.on_conversation_added.assert_called_once_with(conv)
-        conv.on_exit.connect.assert_called_once_with(
-            unittest.mock.ANY
+        self.listener.on_conversation_added.assert_called_once_with(
+            conv,
         )
+
+    def test__add_conversation_adds_conversation_to_conversations(self):
+        conv = unittest.mock.Mock()
+        self.s._add_conversation(conv)
 
         self.assertCountEqual(
             self.s.conversations,
@@ -64,10 +63,54 @@ class TestConversationService(unittest.TestCase):
             ]
         )
 
-        (_, (cb, ), _), = conv.on_exit.mock_calls
+    def test_on_conversation_added_sees_added_conversation(self):
+        captured_conversations = None
+
+        def cb(conv):
+            nonlocal captured_conversations
+            captured_conversations = list(self.s.conversations)
+
+        conv = unittest.mock.Mock()
+        self.s.on_conversation_added.connect(cb)
+        self.s._add_conversation(conv)
+
+        self.assertIn(
+            conv,
+            captured_conversations,
+        )
+
+    def test_removes_added_conversation_on_exit(self):
+        conv = unittest.mock.Mock()
+        self.s._add_conversation(conv)
+
+        conv.on_exit.connect.assert_called_once_with(
+            unittest.mock.ANY
+        )
+
+        _, (cb, ), _ = conv.on_exit.mock_calls[-1]
 
         # should ignore its arguments
         cb(unittest.mock.sentinel.foo, bar=unittest.mock.sentinel.fnord)
+
+        self.assertCountEqual(
+            self.s.conversations,
+            [
+            ]
+        )
+
+    def test_removes_added_conversation_on_failure(self):
+        conv = unittest.mock.Mock()
+        self.s._add_conversation(conv)
+
+        conv.on_failure.connect.assert_called_once_with(
+            unittest.mock.ANY
+        )
+
+        _, (cb, ), _ = conv.on_failure.mock_calls[-1]
+
+        # should ignore its arguments
+        cb(unittest.mock.sentinel.foo, unittest.mock.sentinel.bar,
+           bar=unittest.mock.sentinel.fnord)
 
         self.assertCountEqual(
             self.s.conversations,