+++ tests/muc/test_service.py
@@ -45,6 +45,7 @@ from aioxmpp.testutils import (
     make_connected_client,
     run_coroutine,
     CoroutineMock,
+    make_listener,
 )
 
 
@@ -1990,14 +1991,14 @@ class TestRoom(unittest.TestCase):
         self.assertEqual(stanza.show, aioxmpp.PresenceShow.NONE)
 
         self.jmuc.on_exit(muc_leave_mode=object(),
-                            muc_actor=object(),
-                            muc_reason=object())
+                          muc_actor=object(),
+                          muc_reason=object())
 
         self.assertIsNone(run_coroutine(fut))
 
         self.jmuc.on_exit(muc_leave_mode=object(),
-                            muc_actor=object(),
-                            muc_reason=object())
+                          muc_actor=object(),
+                          muc_reason=object())
 
     def test_members(self):
         presence = aioxmpp.stanza.Presence(
@@ -2924,6 +2925,12 @@ class TestService(unittest.TestCase):
             service.Service
         ))
 
+    def test_is_conversation_service(self):
+        self.assertTrue(issubclass(
+            muc_service.MUCClient,
+            im_conversation.AbstractConversationService,
+        ))
+
     def setUp(self):
         self.cc = make_connected_client()
         self.im_dispatcher = im_dispatcher.IMDispatcher(self.cc)
@@ -2938,6 +2945,7 @@ class TestService(unittest.TestCase):
             im_service.ConversationService: self.im_service,
             aioxmpp.tracking.BasicTrackingService: self.tracking_service,
         })
+        self.listener = make_listener(self.s)
 
     def test_depends_on_IMDispatcher(self):
         self.assertIn(
@@ -3115,6 +3123,7 @@ class TestService(unittest.TestCase):
         room, future = self.s.join(TEST_MUC_JID, "thirdwitch")
 
         self.im_service._add_conversation.assert_called_once_with(room)
+        self.listener.on_conversation_new.assert_called_once_with(room)
 
         self.assertIs(
             self.s.get_muc(TEST_MUC_JID),
@@ -3320,9 +3329,46 @@ class TestService(unittest.TestCase):
         with self.assertRaises(KeyError):
             self.s.get_muc(TEST_MUC_JID)
 
+    def test_on_failure_is_emitted_on_join_error(self):
+        room, future = self.s.join(TEST_MUC_JID, "thirdwitch")
+        listener = make_listener(room)
+
+        response = aioxmpp.stanza.Presence(
+            from_=TEST_MUC_JID,
+            type_=aioxmpp.structs.PresenceType.ERROR)
+        response.error = aioxmpp.stanza.Error()
+        self.s._handle_presence(
+            response,
+            response.from_,
+            False,
+        )
+
+        run_coroutine(asyncio.sleep(0))
+
+        listener.on_enter.assert_not_called()
+        listener.on_failure.assert_called_once_with(
+            future.exception(),
+        )
+
+    def test_on_failure_is_emitted_on_stream_destruction_without_autorejoin(
+            self):
+        room, future = self.s.join(TEST_MUC_JID, "thirdwitch",
+                                   autorejoin=False)
+        listener = make_listener(room)
+
+        self.s._stream_destroyed()
+
+        run_coroutine(asyncio.sleep(0))
+
+        listener.on_enter.assert_not_called()
+        listener.on_failure.assert_called_once_with(
+            future.exception(),
+        )
+
     def test_pending_muc_removed_and_unavailable_presence_emitted_on_cancel(
             self):
         room, future = self.s.join(TEST_MUC_JID, "thirdwitch")
+        listener = make_listener(room)
 
         self.cc.stream.enqueue.mock_calls.clear()
 
@@ -3346,6 +3392,12 @@ class TestService(unittest.TestCase):
             stanza.xep0045_muc,
             muc_xso.GenericExt
         )
+
+        listener.on_failure.assert_called_once_with(unittest.mock.ANY)
+
+        _, (exc,), _ = listener.on_failure.mock_calls[-1]
+
+        self.assertIsInstance(exc, asyncio.CancelledError)
 
     def test_join_completed_on_self_presence(self):
         room, future = self.s.join(TEST_MUC_JID, "thirdwitch")